Second Update
Fully functional.  Calls on an API to receive the information of 12 randomly generated employees and then using JQuery dynamically adds 'employee cards' for each of those 12 employees to the web page. When one of these 'employee cards' is clicked, it brings up a modal window with more detailed information for that employee.